Fixing subtle bug in scale and delete operations
The scale and delete operations both needed to check
for the newest replication controller for the given deployment.
In the event of a config change, the replication controller also gets
updated and changes names.  In order to adapt for this, the scale
and delete operations check for the newest replication controller
for the given deployment.  Also removing console.info messages
that are not necessary.